% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_exported.R
\docType{class}
\name{common_stanfit-class}
\alias{common_stanfit}
\alias{common_stanfit-class}
\title{The common Reference Class for edstan.}
\description{
The common Reference Class for edstan.
}
\section{Fields}{

\describe{
\item{\code{fit}}{A \code{stanfit} object.}

\item{\code{data}}{The data list used in stan model fitting.}
}}
\section{Methods}{

\describe{
\item{\code{autocor(pars, ...)}}{Plot auto-correlations for parameter draws.}

\item{\code{pairs(...)}}{Plot pairs of posterior parameter draws.}

\item{\code{parameters(...)}}{Retrieve a data frame of parameter estimates.}

\item{\code{plot(...)}}{Plot posterior credible intervals.}

\item{\code{posterior(pars, ...)}}{Plot posterior density for parameter draws.}

\item{\code{stancode()}}{Return the Stan code}

\item{\code{traceplot(pars, ...)}}{Draw traceplots.}
}}
\seealso{
The \code{common_stanfit} methods are closely related to \code{rstan} and
\code{edstan} functions. See \code{rstan} functions
\code{\link[rstan]{pairs.stanfit}}, \code{\link[rstan]{plot-method}},
\code{\link[rstan]{traceplot}}, and \code{\link[rstan]{get_stancode}}. See
\code{edstan} functions \code{\link{get_parameters}} and
\code{\link{plot_autocor}}.
}

