% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{convergence_plot}
\alias{convergence_plot}
\title{Plot convergence diagnostics}
\usage{
convergence_plot(
  fit,
  include = NA,
  exclude = "y_rep",
  stats = c("n_eff", "Rhat")
)
}
\description{
Plot convergence diagnostics
}
