% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{.long_format_covariates}
\alias{.long_format_covariates}
\title{Convert covariate data frame for long format data}
\usage{
.long_format_covariates(covariates, jj, formula)
}
\arguments{
\item{covariates}{A data frame containing covariates.}

\item{jj}{Index for person associated with each row.}

\item{not_missing}{Boolean for whether each row is associated with a missing
item response.}
}
\value{
A data frame with one row per person.
}
\description{
Intended for internal use only.
}
\keyword{internal}
