% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{.validate_binary}
\alias{.validate_binary}
\title{Validate a boolean covariate}
\usage{
.validate_binary(x, nm)
}
\arguments{
\item{x}{A vector of covariate values.}

\item{nm}{Name for the covariate.}
}
\value{
A character vector of identified issues.
}
\description{
Intended for internal use only.
}
\keyword{internal}
