% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{.validate_continuous}
\alias{.validate_continuous}
\title{Validate a continuous covariate}
\usage{
.validate_continuous(x, nm)
}
\arguments{
\item{x}{A vector of covariate values.}

\item{nm}{Name for the covariate.}
}
\value{
A character vector of identified issues.
}
\description{
Intended for internal use only.
}
\keyword{internal}
