% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{.validate_regression_model}
\alias{.validate_regression_model}
\title{Validate formula and covariate data}
\usage{
.validate_regression_model(formula, data)
}
\arguments{
\item{formula}{A formula for the latent regression.}

\item{data}{A data frame of covariates.}
}
\value{
A character vector of identified issues.
}
\description{
Intended for internal use only.
}
\keyword{internal}
