% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{edstan_model_code}
\alias{edstan_model_code}
\title{Read and print the code for an edstan model}
\usage{
edstan_model_code(filename, print = TRUE)
}
\arguments{
\item{filename}{The name of the stan file.}

\item{print}{Whether to print the stan file contents. Default is `TRUE`.}
}
\value{
Invisibly returns a character vector of the stan file contents.
}
\description{
This function reads a Stan file from the `inst/extdata/` directory of the
package, returning its contents invisibly while optionally printing them.
}
\examples{
# View the Stan code for the Rasch model
edstan_model_code("rasch_latent_reg.stan")
}
