% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-edstanfit.R
\docType{class}
\name{edstanfit-class}
\alias{edstanfit-class}
\alias{edstanfit}
\title{An S4 class for a fitted edstan model.}
\description{
An S4 class for a fitted edstan model.
}
\section{Slots}{

\describe{
\item{\code{edstan_options}}{A list of options provided when fitting the model}
}}

