% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ability.R
\name{extract_ability}
\alias{extract_ability}
\title{Extract ability estimates}
\usage{
extract_ability(object, probs = c(0.1, 0.9), rescale_by_alpha = FALSE)
}
\description{
Extract ability estimates
}
