% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_common.R
\name{get_best_logLik}
\alias{get_best_logLik}
\title{Get the posterior log-likelihood evaluated at posterior parameter means.}
\usage{
get_best_logLik(fit, data)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{data}{The data list provided in to stan.}
}
\value{
The posterior log-likelihood evaluated at posterior parameter means.
}
\description{
Get the posterior log-likelihood evaluated at posterior parameter means.
}
\examples{
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
ll <- get_best_logLik(stan_fit)
}

