% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_common.R
\name{get_dic}
\alias{get_dic}
\title{View mean deviance, deviance evaluated at posterior parameter means, and DIC.}
\usage{
get_dic(fit, data)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{data}{The data list provided in to stan.}
}
\value{
The posterior log-likelihood evaluated at posterior parameter means.
}
\description{
View mean deviance, deviance evaluated at posterior parameter means, and DIC.
}
\examples{
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
get_dic(stan_fit)
}

