% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_common.R
\name{get_mean_logLik}
\alias{get_mean_logLik}
\title{Get the mean posterior log-likelihood from a \code{stanfit} object.}
\usage{
get_mean_logLik(fit)
}
\arguments{
\item{fit}{A \code{stanfit} object.}
}
\value{
The mean posterior log-likelihood.
}
\description{
Get the mean posterior log-likelihood from a \code{stanfit} object.
}
\examples{
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
ll <- get_mean_logLik(stan_fit)
}

