% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/common_exported.R
\name{get_model_stan}
\alias{get_model_stan}
\title{Load or save a stan model from a .Rdata file.}
\usage{
get_model_stan(model)
}
\arguments{
\item{model}{A string for the name of an \code{edstan} model.}
}
\value{
Returns a Stan model object for the specified model..
}
\description{
Load or save a stan model from a .Rdata file.
}
\details{
This function is used to create a Stan model object from a Stan program file
and then to save the Stan model object in an .Rdata file. The usual use for
this function is create the .Rdata file the first time a given
\code{edstan} wrapper is called. On subsequent calls to the given wrapper,
the Stan model object is just loaded from the .Rdata file, improving the
speed of the estimation. In general, users will have no need to use this
function directly.
}
\examples{
rasch_model_obj <- get_model_stan("rasch")

# If we wanted to call this for every edstan model at once while we take
# a lunch break:
folder <- system.file("extdata", package = "edstan")
files <- dir(folder)
for(f in files) {
  if(grepl("[.]stan$", f)) {
    model <- sub("[.]stan$", "", f)
    get_model_stan(model)
  }
}
}

