% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_exported.R
\name{get_parameters}
\alias{get_parameters}
\title{Retrieve a data frame of parameter estimates from a \code{stanfit} object.}
\usage{
get_parameters(fit, ...)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{...}{Additional parameters passed to \code{\link[rstan]{summary}}.}
}
\value{
A data frame.
}
\description{
Retrieve a data frame of parameter estimates from a \code{stanfit} object.
}
\examples{
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
all_parameters <- get_parameters(stan_fit)
just_thetas <- get_parameters(stan_fit, pars = "theta")
}

