% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_mrasch.R
\name{mrasch_long_stan}
\alias{mrasch_long_stan}
\title{Estimate the multidimensional Rasch model using long-form data.}
\usage{
mrasch_long_stan(id, item, response, z, ...)
}
\arguments{
\item{id}{A vector identifying persons.}

\item{item}{A vector identifying persons.}

\item{response}{A vector coded as 1 for a correct response and 0 otherwise.}

\item{z}{A design matrix for the random effects.}

\item{...}{Additional options passed to \code{\link[rstan]{stan}}.}
}
\value{
A \code{\link{mrasch_stanfit}} object.
}
\description{
Estimate the multidimensional Rasch model using long-form data.
}
\seealso{
See \code{\link{mrasch_wide_stan}} for wide-form data. See \code{\link{mrasch_stanfit}} and \code{\link{common_stanfit}} for applicable methods.
}

