% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_mrasch.R
\name{mrasch_stan}
\alias{mrasch_stan}
\title{Estimate the multidimensional Rasch model using Stan}
\usage{
mrasch_stan(response_matrix = NULL, id = NULL, item = NULL,
  response = NULL, z, ...)
}
\arguments{
\item{response_matrix}{A response matrix. Columns represent items, and rows
represent persons. Each element indicates a correct response by one or
zero, and NA may be supplied for missing responses.}

\item{id}{A vector identifying persons. Required if \code{response_matrix} is
not supplied.}

\item{item}{A vector identifying persons. Required if \code{response_matrix}
is not supplied.}

\item{response}{A vector coded as 1 for a correct response and 0 otherwise.
Required if \code{response_matrix} is not supplied.}

\item{z}{A design matrix for the random effects. It must have one row per
item and one column per dimension.}

\item{...}{Additional options passed to \code{\link[rstan]{sampling}}. The
usual choices are \code{iter} for the number of iterations and
\code{chains} for the number of chains.}
}
\value{
A \code{\link{rasch_stanfit}} Reference Class object.
}
\description{
Estimate the multidimensional Rasch model using Stan
}
\examples{
The spelling data response matrix
x <- spelling[, 2:5]

# We pretend that the first two items in the spelling data load on a different
# dimension than the second two.
z <- cbind(c(1, 1, 0, 0), c(0, 0, 1, 1))

# We estimate the model
myfit <- mrasch(response_matrix = x, z = z, iter = 200, chains = 4)
}
\seealso{
See \code{\link[rstan]{sampling}} for additional options. See
  \code{\link{mrasch_stanfit}} and \code{\link{common_stanfit}} for
  applicable Reference Class methods.
}

