% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_mrasch.R
\docType{class}
\name{mrasch_stanfit-class}
\alias{mrasch_stanfit}
\alias{mrasch_stanfit-class}
\title{The multidimensional Rasch model Reference Class.}
\description{
The multidimensional Rasch model Reference Class.
}
\section{Fields}{

\describe{
\item{\code{person_names}}{A character vector for person names.}

\item{\code{item_names}}{A character vector for item names.}
}}
\section{Methods}{

\describe{
\item{\code{icc(item, ...)}}{Plot an item characteristic curve.}
}}
\seealso{
See \code{\link{common_stanfit}} for additional methods. See
  \code{\link{mrasch_stan}} and for estimating the Rasch model. See
  \code{\link{plot_icc}} for addtional options for \code{icc}.
}

