% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_mrasch.R
\name{mrasch_wide_stan}
\alias{mrasch_wide_stan}
\title{Estimate the multidimensional Rasch model using a response matrix.}
\usage{
mrasch_wide_stan(response_matrix, z, ...)
}
\arguments{
\item{response_matrix}{A response matrix. Columns represent items, and rows represent persons. Each element is one or zero or may be NA if missing.}

\item{z}{A design matrix for the random effects.}

\item{...}{Additional options passed to \code{\link[rstan]{stan}}.}
}
\value{
A \code{\link{mrasch_stanfit}} object.
}
\description{
Estimate the multidimensional Rasch model using a response matrix.
}
\seealso{
See \code{\link{mrasch_long_stan}} for long-form data. See \code{\link{mrasch_stanfit}} and \code{\link{common_stanfit}} for applicable methods.
}

