% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/common_exported.R
\name{plot_autocor}
\alias{plot_autocor}
\title{Plot (or get a table of) auto-correlations for parameter draws.}
\usage{
plot_autocor(fit, pars, back = 10, show_matrix = FALSE,
  show_plot = !show_matrix)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{pars}{Parameters for which to calculate auto-correlations.}

\item{back}{Number of iterations back to calculate auto-correlations. Default is 10.}

\item{show_matrix}{Whether the correlations should be returned as a matrix. Default is FALSE.}

\item{show_plot}{Whether to draw the plots. Default is TRUE.}
}
\value{
If \code{show_plot = TRUE}, a matrix of plots of the auto-correlations for
the chosen parameters. If \code{show_table = TRUE}, the auto-correlations in
the form of a matrix. Both may be requested.
}
\description{
Plot (or get a table of) auto-correlations for parameter draws.
}
\examples{
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
plot_autocor(stan_fit, pars = "alpha")
autocorrs <- plot_autocor(stan_fit, pars = c("alpha", "beta"), show_matrix = TRUE)
}

