% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/common_exported.R
\name{plot_icc}
\alias{plot_icc}
\title{Plot an item characteristic curve.}
\usage{
plot_icc(fit = NULL, item = NULL, new = TRUE, alpha = NULL,
  beta = NULL, gamma = NULL, theta = seq(from = -2, to = 2, length.out =
  100), ...)
}
\arguments{
\item{fit}{A \code{stanfit} object. Required if \code{alpha}, \code{beta}, or \code{gamma} is not specificied.}

\item{item}{Either a single integer for the item to be looked up \code{item_names}. Required if \code{alpha}, \code{beta}, or \code{gamma} is not specificied.}

\item{new}{Indicates whether a new plot should be made. Otherwise, the item characteristic curve is drawn on the active plot. Default is TRUE.}

\item{alpha}{A value for the discrimination parameter. Required if \code{fit} is not supplied.}

\item{beta}{A value for the difficulty parameter. Required if \code{fit} is not supplied.}

\item{gamma}{A value for the pseudo-guessing parameter. Required if \code{fit} is not supplied.}

\item{theta}{A vector of abilities to plot.}

\item{...}{Additional options passed to \code{\link{plot}} or \code{\link{lines}}.}
}
\value{
A single plotted item characteristic curve. If any parameter \code{alpha},
\code{beta}, or \code{gamma} is not supplied, the function attempts to look
up the values using the \code{stanfit} supplied to the \code{fit} option. This look up
requires an item number specified in \code{item}.
}
\description{
Plot an item characteristic curve.
}
\examples{
plot_icc(alpha = 1, beta = 0, gamma = 0)
plot_icc(alpha = 1.2, beta = -1, gamma = .2)

myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit
items <- myfit$item_names
plot_icc(fit = stan_fit, item = 1, gamma = 0)
}

