% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_exported.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot kernal densities for parameter posteriors.}
\usage{
plot_posterior(fit, pars = NULL, overall = TRUE, chains = TRUE,
  legend = FALSE, inc_warmup = FALSE)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{pars}{Parameters for which to plot densities.}

\item{overall}{Whether to plot the overall density. Default is TRUE.}

\item{chains}{Whether to plot densities for individual chains. Default is TRUE.}

\item{legend}{Whether to include a legend. Default is FALSE.}

\item{inc_warmup}{Whether to include warmup draws. Default is FALSE.}
}
\value{
One kernal density plot per parameter.
}
\description{
Plot kernal densities for parameter posteriors.
}
\examples{
# Get an example stanfit object.
myfit <- twopl_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
stan_fit <- myfit$fit

# Plot first discrimination parameter with a legend.
plot_posterior(stan_fit, pars = "alpha[1]", legend = TRUE)

# Show plots for all discrimation parameters on same display.
plot_posterior(stan_fit, pars = "alpha")
}

