% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_rasch.R
\name{rasch_long_stan}
\alias{rasch_long_stan}
\title{Estimate the Rasch model using long-form data.}
\usage{
rasch_long_stan(id, item, response, ...)
}
\arguments{
\item{id}{A vector identifying persons.}

\item{item}{A vector identifying persons.}

\item{response}{A vector coded as 1 for a correct response and 0 otherwise.}

\item{...}{Additional options passed to \code{\link[rstan]{stan}}.}
}
\value{
A \code{\link{rasch_stanfit}} object.
}
\description{
Estimate the Rasch model using long-form data.
}
\examples{
# Make the spelling data long-form
require(reshape2)
wide <- spelling[, 2:5]
wide$id <- 1:nrow(wide)
long <- melt(wide, id.vars = "id", variable.name = "item", value.name = "response")

# Estimate the model
myfit <- rasch_long_stan(long$id, long$item, long$response, chains = 4, iter = 200)
myfit$show()
}
\seealso{
See \code{\link{rasch_wide_stan}} for wide-form data. See \code{\link{rasch_stanfit}} and \code{\link{common_stanfit}} for applicable methods.
}

