% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_rasch.R
\docType{class}
\name{rasch_stanfit-class}
\alias{rasch_stanfit}
\alias{rasch_stanfit-class}
\title{The Rasch model Reference Class}
\description{
The Rasch model Reference Class
}
\section{Fields}{

\describe{
\item{\code{person_names}}{A character vector for person names.}

\item{\code{item_names}}{A character vector for item names.}
}}
\section{Methods}{

\describe{
\item{\code{icc(item, ...)}}{Plot an item characteristic curve.}
}}
\seealso{
See \code{\link{common_stanfit}} for additional methods. See
  \code{\link{rasch_long_stan}} and \code{\link{rasch_wide_stan}} for
  estimating the Rasch model. See \code{\link{plot_icc}} for addtional
  options for \code{icc}.
}

