% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_rasch.R
\name{rasch_wide_stan}
\alias{rasch_wide_stan}
\title{Estimate the Rasch model using a response matrix.}
\usage{
rasch_wide_stan(response_matrix, ...)
}
\arguments{
\item{response_matrix}{A response matrix. Columns represent items, and rows represent persons. Each element is one or zero or may be NA if missing.}

\item{...}{Additional options passed to \code{\link[rstan]{sampling}}.}
}
\value{
A \code{\link{rasch_stanfit}} object.
}
\description{
Estimate the Rasch model using a response matrix.
}
\examples{
myfit <- rasch_wide_stan(spelling[, 2:5], chains = 4, iter = 200)
myfit$show()
}
\seealso{
See \code{\link{rasch_long_stan}} for long-form data. See \code{\link{rasch_stanfit}} and \code{\link{common_stanfit}} for applicable methods.
}

