% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{rhat_columns}
\alias{rhat_columns}
\title{View a plot of Rhat statistics after using \code{irt_stan}}
\usage{
rhat_columns(fit, ...)
}
\arguments{
\item{fit}{A \code{stanfit-class} object created by \code{\link{irt_stan}}
or \code{\link[rstan]{stan}}.}

\item{...}{Additional options (such as \code{pars} or \code{use_cache}),
passed to the \code{summary} method for a \code{stanfit} object. Not
required.}
}
\value{
A \code{ggplot} object.
}
\description{
View a plot of Rhat statistics after using \code{irt_stan}
}
\examples{
# Make a suitable data list:
spelling_list <- irt_data(response_matrix = spelling[, 2:5],
                          covariates = spelling[, "male", drop = FALSE],
                          formula = ~ 1 + male)

\dontrun{
# Fit a latent regression  2PL
twopl_fit <- irt_stan(spelling_list, model = "2pl_latent_reg.stan",
                      iter = 200, chains = 4)

# Get a plot of Rhat statistics
rhat_columns(twopl_fit)
}
}
\seealso{
See \code{\link[rstan]{stan_rhat}}, which provides a histogram of
  Rhat statistics.
}

