\name{spelling}
\docType{data}
\alias{spelling}
\title{Spelling data}
\description{Item response data regarding  student spelling performance on four words: \emph{infidelity}, \emph{panoramic}, \emph{succumb}, and \emph{girder}. The sample includes 284 male and 374 female undergraduate students from the University of Kansas. Each item was scored as either correct or incorrect.}
\usage{spelling}
\format{
A wide-form data.frame (one row per person) with the following columns:
	\describe{
  \item{male}{Indicator for whether person is male.}
  \item{infidelity}{Indicator for whether person spelled \emph{infidelity} correctly.}
  \item{panoramic}{Indicator for whether person spelled \emph{panoramic} correctly.}
  \item{succumb}{Indicator for whether person spelled \emph{succumb} correctly.}
  \item{girder}{Indicator for whether person spelled \emph{girder} correctly.}
  }
}
\source{Thissen, D., Steinberg, L. and Wainer, H. (1993). Detection of Differential Item Functioning Using the Parameters of Item Response Models. In \emph{Differential Item Functioning}, edited by Holland. P. and Wainer, H., 67-114. Hillsdale, NJ: Lawrence Erlbaum.}
\keyword{datasets}
