% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_twopl.R
\name{twopl_long_stan}
\alias{twopl_long_stan}
\title{Estimate the 2PL model using long-form data.}
\usage{
twopl_long_stan(id, item, response, path = "", ...)
}
\arguments{
\item{id}{A vector identifying persons.}

\item{item}{A vector identifying persons.}

\item{response}{A vector coded as 1 for a correct response and 0 otherwise.}

\item{...}{Additional options passed to stan().}
}
\value{
A twopl_stanfit reference class object.
}
\description{
Estimate the 2PL model using long-form data.
}
\examples{
require(reshape2)
wide <- cito[, 2:19]
wide$id <- 1:nrow(wide)
long <- melt(wide, id.vars = "id", variable_name = "item")
twopl_long_stan(long$id, long$item, long$value, chains = 4, iter = 200)
}

