% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_twopl.R
\docType{class}
\name{twopl_stanfit-class}
\alias{twopl_stanfit}
\alias{twopl_stanfit-class}
\title{The two-parameteric logistic Reference Class.}
\description{
The two-parameteric logistic Reference Class.
}
\section{Fields}{

\describe{
\item{\code{person_names}}{A character vector for person names.}

\item{\code{item_names}}{A character vector for item names.}
}}
\section{Methods}{

\describe{
\item{\code{icc(item, ...)}}{Plot an item characteristic curve.}
}}
\seealso{
See \code{\link{common_stanfit}} for additional methods. See
  \code{\link{twopl_stan}} for estimating the 2PL. See \code{\link{plot_icc}}
  for addtional options for \code{icc}.
}

