% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_twopl.R
\name{twopl_wide_stan}
\alias{twopl_wide_stan}
\title{Estimate the 2PL model using a response matrix.}
\usage{
twopl_wide_stan(response_matrix, ...)
}
\arguments{
\item{response_matrix}{A response matrix. Columns represent items, and rows represent persons. Each element is one or zero or may be NA if missing.}

\item{...}{Additional options passed to stan().}
}
\value{
A twopl_stanfit reference class object.
}
\description{
Estimate the 2PL model using a response matrix.
}
\examples{
myfit <- twopl_wide_stan(cito[, 2:19], chains = 4, iter = 200)
myfit$results()
}

