% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper_common.R
\docType{class}
\name{common_stanfit-class}
\alias{common_stanfit}
\alias{common_stanfit-class}
\title{The common Reference Class for edstan.}
\description{
The common Reference Class for edstan.
}
\section{Fields}{

\describe{
\item{\code{fit}}{A \code{stanfit} object.}

\item{\code{data}}{The data list used in stan model fitting.}
}}
\section{Methods}{

\describe{
\item{\code{autocor(pars, ...)}}{Plot auto-correlations for parameter draws.}

\item{\code{best_logLik()}}{Get the posterior log-likelihood evaluated at posterior parameter means.}

\item{\code{dic()}}{View mean deviance, deviance evaluated at posterior parameter means, and DIC.}

\item{\code{mean_logLik()}}{Get the mean posterior log-likelihood.}

\item{\code{pairs(...)}}{Plot pairs of posterior parameter draws.}

\item{\code{parameters(...)}}{Retrieve a data frame of parameter estimates.}

\item{\code{plot(...)}}{Plot posterior credible intervals.}

\item{\code{traceplot(...)}}{Draw traceplots.}
}}
\seealso{
The \code{common_stanfit} methods are closely related to \code{rstan} and \code{edstan} functions.
See \code{rstan} functions \code{\link[rstan]{pairs.stanfit}},
\code{\link[rstan]{plot-method}}, and \code{\link[rstan]{traceplot}}.
See \code{edstan} functions \code{\link{get_parameters}},
\code{\link{get_mean_logLik}}, \code{\link{get_best_logLik}},
\code{\link{get_dic}}, and \code{\link{plot_autocor}}.
}

