% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{irt_data}
\alias{irt_data}
\title{Create a Stan data list from an item response matrix or from long-form data}
\usage{
irt_data(response_matrix = matrix(), y = integer(), ii = integer(),
  jj = integer(), W = matrix())
}
\arguments{
\item{response_matrix}{An item response matrix.
Columns represent items, and rows represent persons.
NA may be supplied for missing responses.
The lowest score for each item should be 0, with exception to rating scale
models.}

\item{y}{A vector of scored responses for long-form data.
The lowest score for each item should be 0, with exception to rating scale
models.
NAs are not permitted, but missing responses may simply be ommitted
instead.
Required if \code{response_matrix} is not supplied.}

\item{ii}{A vector indexing the items in \code{y}.
This must consist of consecutive integers starting at 1.
\code{\link{labelled_integer}} may be used to create a suitable vector.
Required if \code{response_matrix} is not supplied.}

\item{jj}{A vector indexing the persons in \code{y}.
This must consist of consecutive integers starting at 1.
\code{\link{labelled_integer}} may be used to create a suitable vector.
Required if \code{response_matrix} is not supplied.}

\item{W}{An optional matrix of person-covariates.}
}
\value{
A data list suitable for \code{\link{irt_stan}}.
}
\description{
Create a Stan data list from an item response matrix or from long-form data
}
\examples{
# If the data are in a response matrix ("wide-form"):
X <- spelling[, 2:5]
W <- spelling[, 1]
data_list <- irt_data(X, W = W)
}
\seealso{
See \code{\link{irt_stan}} to fit a model to the data list.
}

