% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{irt_stan}
\alias{irt_stan}
\title{Estimate an item response model with Stan}
\usage{
irt_stan(data_list, model = "", ...)
}
\arguments{
\item{data_list}{A Stan data list created with \code{\link{irt_data}}.}

\item{model}{The file name for one of the provided .stan files, or
alternatively, a user-created .stan file that accepts \code{data_list} as
input data.
The ".stan" file extension may be omitted.
Defaults to either rasch_latent_reg.stan or pcm_latent_reg.stan.}

\item{...}{Additional options passed to \code{\link[rstan]{stan}}. The
usual choices are \code{iter} for the number of iterations and
\code{chains} for the number of chains.}
}
\value{
A \code{\link[rstan]{stanfit}} object.
}
\description{
Estimate an item response model with Stan
}
\examples{
X <- spelling[, 2:5]
W <- spelling[, 1]
data_list <- irt_data(X, W = W)

}
\seealso{
This function is a wrapper for \code{\link[rstan]{stan}}.
  See \code{\link{irt_data}} and \code{\link{labelled_integer}} for functions
  that facilitate creating a suitable \code{data_list}.
  See \code{\link{print_irt_stan}} and \code{\link[rstan]{print}} for ways of
  getting tables summarizing parameter posteriors.
}

