% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_stan.R
\name{irt_stan}
\alias{irt_stan}
\title{Fit an IRT model}
\usage{
irt_stan(
  response_matrix = matrix(),
  item = integer(),
  person = integer(),
  response = integer(),
  varying_slopes = TRUE,
  thresholds = TRUE,
  common_steps = FALSE,
  posterior_replicates = FALSE,
  covariates = NULL,
  prior_alpha = if (varying_slopes) c(mean = 0, sd = 1) else c(mean = 0, sd = 1),
  prior_beta_base = c(mean = 0, sd = 3),
  prior_beta_step = if (thresholds) c(mean = 0.5, sd = 1) else c(mean = 0, sd = 3),
  prior_lambda = NULL,
  return_data_list = FALSE,
  ...
)
}
\description{
Fit an IRT model
}
