% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_stan.R
\name{irt_stan}
\alias{irt_stan}
\title{Fit an IRT model}
\usage{
irt_stan(
  item,
  person,
  y,
  varying_slopes = TRUE,
  thresholds = TRUE,
  common_steps = FALSE,
  posterior_replicates = FALSE,
  covariates = NULL,
  prior_alpha = if (varying_slopes) c(0, 1) else c(0, 1),
  prior_beta_base = c(0, 3),
  prior_beta_step = if (thresholds) c(0.5, 1) else c(0, 3),
  prior_lambda = NULL,
  return_data_list = FALSE,
  ...
)
}
\description{
Fit an IRT model
}
